adjustments integer to long

+++ core/src/main/java/br/com/bus/application/service/ItinerarioService.java
@@ -26,12 +26,12 @@ public class ItinerarioService {
     DeletarItinerario deletar;
 
     public List<ItinerarioDTO> listar(int page, int size) { return busca.listar(page, size); }
-    public ItinerarioDTO buscarPorId(Integer idLinha, Integer idPontoParada) {
+    public ItinerarioDTO buscarPorId(Long idLinha, Long idPontoParada) {
         return busca.porId(idLinha, idPontoParada);
     }
     public ItinerarioDTO criar(ItinerarioDTO dto) { return criar.executar(dto); }
-    public ItinerarioDTO atualizar(Integer idLinha, Integer idPontoParada, ItinerarioDTO dto) {
+    public ItinerarioDTO atualizar(Long idLinha, Long idPontoParada, ItinerarioDTO dto) {
         return atualizar.executar(idLinha, idPontoParada, dto);
     }
-    public void deletar(Integer idLinha, Integer idPontoParada) { deletar.executar(idLinha, idPontoParada); }
+    public void deletar(Long idLinha, Long idPontoParada) { deletar.executar(idLinha, idPontoParada); }
 }
